+++ CHANGELOG.md
@@ -66,6 +66,13 @@
 
 ## Unreleased
 
+### Additions
+
+#### Performance
+
+- Data plane's connection to control plane is moved to a privileged worker process
+  [#9564](https://github.com/Kong/kong/pull/9564)
+
 ### Fixes
 
 #### Admin API